+++ compiled/vwf.h
@@ -13,7 +13,7 @@ byte decode_character(byte chr);
 int get_tile_number(int x, int y);
 int expand_bit_depth(byte row, int foreground);
 byte reduce_bit_depth(int row, int foreground);
-byte print_character(byte chr, byte x, byte y, byte font, byte foreground);
+byte print_character(byte chr, int x, int y, int font, int foreground);
 byte print_character_with_callback(byte chr, int x, int y, int font, int foreground,
     int *dest, int (*getTileCallback)(int, int), int useTilemap);
 byte print_character_to_ram(byte chr, int *dest, int xOffset, int font, int foreground);
